mesh: Extract protocol and framing from mesh-io-silvair
